# prop/portal/openapi/prop-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Prop.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portal/password-reset": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/portal/password-reset/check": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/password-reset/confirm": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalMeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/admin/password-reset": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/portal/admin/password-reset/check": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/admin/password-reset/confirm": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalMeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/invites/check": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/admin/invites/check": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCheckResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "required": [
          "token",
          "password"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LinkCheckRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "The token from a link in an email."
      },
      "LinkCheckResponse": {
        "required": [
          "status",
          "email",
          "hasPassword"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LinkStatus"
          },
          "email": {
            "type": "string"
          },
          "hasPassword": {
            "type": "boolean"
          }
        },
        "description": "Whether a link from an email still works and whose it is. HasPassword tells whether the person can\nalready log in, for example after using the invitation once."
      },
      "LinkStatus": {
        "enum": [
          "Valid",
          "Used",
          "Expired"
        ],
        "description": "What a one-time link from an email is worth when it is opened."
      },
...
      "PasswordResetRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "The email of the person who forgot the password."
      },
...
      "PortalMeResponse": {
        "required": [
          "userId",
          "email",
          "role",
          "firmName",
          "emailConfirmed",
          "name",
          "country"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "firmName": {
            "type": "string"
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Who is logged in to the portal. Role is trader or admin. EmailConfirmed is false\nfor a trader who chose a password on an order's page and has not opened the link from the email since.\nName and Country are a trader's from buying, so the shop need not ask again."
      },
...
    }
  },
...
}